+++ ChangeLog.md
@@ -1,4 +1,8 @@
-# 0.2.0 [2022-06-16]
+# 0.3.0 [2022-06-16]
+- FIX for duplicate entries in project file
+- add feature PackageReleaseNotes
+- unify unique project/tool name "KsWare.BuildTools.IncrementVersion"
+# 0.2.0
 - improvements
 # 0.1.1
 - build logic improved